fix(vk-wrapper): reorganize parameter order in Command create method and resize command buffers
- Reorganized the parameter order in the `Command::create` method to improve readability and consistency.
- Resized the command buffers to a fixed size of `MAX_FRAMES_IN_FLIGHT` to optimize memory usage.

+++ src/plugin/vk-wrapper/imageView/ImageView.cpp
@@ -15,7 +15,7 @@ void ImageView::create(const VkDevice device, std::vector<VkImage> swapChainImag
 {
     _swapChainImageViews.resize(swapChainImages.size());
 
-    for (size_t i = 0; i < swapChainImages.size(); ++i)
+    for (uint32_t i = 0; i < swapChainImages.size(); ++i)
     {
         VkImageViewCreateInfo createInfo{};
         createInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
